Fix a bunch of imports for the renaming SQS/SNS code
