+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/services/TransformerWorkerService.scala
@@ -2,9 +2,9 @@ package uk.ac.wellcome.transformer.services
 
 import akka.actor.ActorSystem
 import com.google.inject.Inject
+import uk.ac.wellcome.messaging.sqs.{SQSReader, SQSWorker}
 import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.aws.SQSMessage
-import uk.ac.wellcome.sqs.{SQSReader, SQSWorker}
 import uk.ac.wellcome.transformer.receive.SQSMessageReceiver
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
