+++ data_api/snapshot_generator/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/services/SnapshotGeneratorWorkerService.scala
@@ -2,11 +2,11 @@ package uk.ac.wellcome.platform.snapshot_generator.services
 
 import akka.actor.ActorSystem
 import com.google.inject.Inject
+import uk.ac.wellcome.messaging.sns.SNSWriter
+import uk.ac.wellcome.messaging.sqs.{SQSReader, SQSWorker}
 import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.aws.SQSMessage
 import uk.ac.wellcome.platform.snapshot_generator.models.SnapshotJob
-import uk.ac.wellcome.sns.SNSWriter
-import uk.ac.wellcome.sqs.{SQSReader, SQSWorker}
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import uk.ac.wellcome.utils.JsonUtil._
