+++ sierra_adapter/sierra_bib_merger/src/main/scala/uk/ac/wellcome/platform/sierra_bib_merger/services/SierraBibMergerWorkerService.scala
@@ -3,10 +3,9 @@ package uk.ac.wellcome.platform.sierra_bib_merger.services
 import akka.actor.ActorSystem
 import com.google.inject.Inject
 import io.circe.Decoder
-import uk.ac.wellcome.sqs.SQSWorkerToDynamo
+import uk.ac.wellcome.messaging.sqs.{SQSReader, SQSWorkerToDynamo}
 import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.transformable.sierra.SierraRecord
-import uk.ac.wellcome.sqs.SQSReader
 import uk.ac.wellcome.utils.JsonUtil._
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
