+++ sierra_adapter/sierra_items_to_dynamo/src/main/scala/uk/ac/wellcome/platform/sierra_items_to_dynamo/services/SierraItemsToDynamoWorkerService.scala
@@ -3,11 +3,11 @@ package uk.ac.wellcome.platform.sierra_items_to_dynamo.services
 import akka.actor.ActorSystem
 import com.google.inject.Inject
 import io.circe.Decoder
-import uk.ac.wellcome.sqs.SQSWorkerToDynamo
+import uk.ac.wellcome.messaging.sqs.SQSWorkerToDynamo
 import uk.ac.wellcome.utils.JsonUtil._
+import uk.ac.wellcome.messaging.sqs.SQSReader
 import uk.ac.wellcome.metrics.MetricsSender
 import uk.ac.wellcome.models.transformable.sierra.SierraRecord
-import uk.ac.wellcome.sqs.SQSReader
 import uk.ac.wellcome.utils.GlobalExecutionContext._
 import io.circe.generic.extras.semiauto._
 import scala.concurrent.Future
